Change schedule and company links to details/summary, divide ContactPageView to 3 separates views
