--- src/components/Header/HomePageLink.module.scss
@@ -1,36 +1,0 @@
-@import "../../variables.scss";
-
-.homePageLink {
-  position: fixed;
-  top: 0;
-  left: 50%;
-  transform: translateX(-50%);
-  width: auto;
-  max-width: 200px;
-  padding: 0 32px;
-  height: 51px;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  z-index: 11;
-  transition: background 0.5s ease-out;
-  cursor: pointer;
-
-  a {
-    color: $base-green;
-    text-decoration: none;
-    text-transform: uppercase;
-    font-size: 16px;
-    letter-spacing: 0.05em;
-    font-weight: 700;
-    font-family: $base-font;
-    transition: color 0.5s ease-out;
-  }
-
-  &:hover {
-    background: $base-green;
-    a {
-      color: $white;
-    }
-  }
-}
