--- src/components/Header/TopMenuItemsSecondary.js
@@ -1,20 +1,0 @@
-import React from "react";
-// import { Link } from "react-router-dom";
-// import styles from "./TopMenuItemsSecondary.module.scss";
-
-class TopMenuItemsSecondary extends React.Component {
-  render() {
-    return (
-      // <Link
-      //   to="/"
-      //   className={styles.topMenuItem}
-      //   activeClass={styles.topMenuItemActive}
-      //   onClick={this.props.openMobileMenu}>
-      //   home
-      // </Link>
-      <></>
-    );
-  }
-}
-
-export default TopMenuItemsSecondary;
