--- src/components/Header/TopMenuItemsSecondary.module.scss
@@ -1,32 +1,0 @@
-@import "../../variables.scss";
-
-.topMenuItem {
-  display: block;
-  height: 26px;
-  margin: 16px 0;
-  cursor: pointer;
-  text-align: center;
-  color: $base-dark;
-  font-size: 18px;
-  line-height: 26px;
-  letter-spacing: 0.05em;
-  font-weight: 700;
-  font-family: $base-font;
-  text-decoration: none;
-}
-
-@media (min-width: 768px) {
-  .topMenuItem {
-    height: 51px;
-    line-height: 51px;
-    font-size: 16px;
-    padding: 0 32px;
-    color: $base-green;
-    transition: color 0.5s ease-out, background 0.5s ease-out;
-
-    &:hover {
-      color: $white;
-      background: $base-green;
-    }
-  }
-}
