--- src/components/Header/TopMenuSecondary.module.scss
@@ -1,123 +1,0 @@
-@import "../../variables.scss";
-
-.hexagonLogoWrapper {
-  display: none;
-
-  @media screen and (min-width: 1230px) {
-    display: block;
-    position: fixed;
-    z-index: 6;
-    top: 0;
-    left: calc((100vw - 1230px - 8px) / 2);
-  }
-}
-
-.hamburgerWrapper {
-  position: fixed;
-  top: 0;
-  left: 0;
-  height: 60px;
-  width: 100vw;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-  padding-left: 20px;
-  background-color: $white;
-  z-index: 1001;
-}
-
-.topMenuWrapperMobileHidden {
-  position: fixed;
-  top: -100vh;
-  right: 0;
-  width: 100%;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  text-transform: uppercase;
-  z-index: 1000;
-  background-color: $white;
-  transition: transform 1s;
-  overflow: hidden;
-
-  .mobileMenuPatternTop {
-    position: absolute;
-    top: 32px;
-    left: 0;
-  }
-
-  .mobileMenuPatternBottom {
-    position: absolute;
-    bottom: -32px;
-    right: 0;
-    transform: rotate(-180deg);
-    transform-origin: center;
-  }
-
-  .additionalItem {
-    margin: 16px 0;
-    .phoneNumber {
-      margin: 16px 0;
-      line-height: 23px;
-      font-size: 16px;
-      font-weight: 700;
-      color: $base-dark;
-      text-decoration: none;
-      letter-spacing: $medium-letter-spacing;
-
-      &:visited {
-        color: $base-dark;
-      }
-    }
-  }
-}
-
-.topMenuWrapperMobileOpen {
-  @extend .topMenuWrapperMobileHidden;
-  transform: translateY(100vh);
-  transition: transform 1s;
-}
-
-.topMenuWrapper {
-  // @extend .topMenuWrapperMobileHidden;
-  position: fixed;
-  top: 0;
-  left: 0;
-  margin: 0;
-  padding: 0;
-  width: 100vw;
-  height: 51px;
-  list-style: none;
-  text-transform: uppercase;
-  display: flex;
-  flex-direction: row;
-  justify-content: center;
-  align-items: center;
-  background: $white;
-  z-index: 5;
-}
-
-// .scrollToTop {
-//   position: fixed;
-//   bottom: 60px;
-//   right: 0;
-//   width: 40px;
-//   height: 40px;
-//   background-color: rgba($color-4, .4);
-//   //border: 1px solid rgba($footer-text-color, .9);
-//   z-index: 1001;
-//   display: flex;
-//   justify-content: center;
-//   align-items: center;
-//   cursor: pointer;
-
-//   .scrollToTopIcon {
-//     color: rgba($footer-text-color, .9);
-//     font-size: 1.5rem;
-//   }
-// }
