+++ src/views/ContactFormView/ContactFormView.module.scss
@@ -1,0 +1,175 @@
+@import "../../variables.scss";
+@import "../../mixins.scss";
+
+.contactPageWrapper {
+  grid-area: hero;
+  .sectionTitle {
+    @include section-title;
+    margin: 140px 0 80px;
+    text-transform: uppercase;
+    text-align: left;
+  }
+
+  .sectionParagraph {
+    @include base-paragraph;
+  }
+
+  .contactFormSection {
+    .sectionSubTitle {
+      @include section-sub-title;
+      font-size: 32px;
+      margin: 96px 0 24px;
+      text-align: left;
+    }
+
+    .sectionParagraph {
+      @include base-paragraph;
+    }
+
+    .contactFormWrapper {
+      margin-top: 64px;
+      display: grid;
+      grid-template-columns: auto 633px;
+      grid-template-rows: auto;
+      grid-template-areas: "left right";
+
+      .contactData {
+        grid-area: left;
+      }
+
+      .contactData {
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+
+        .logo {
+          margin-bottom: 60px;
+        }
+
+        .adressWrapper,
+        .socialWrapper {
+          margin: 32px 0;
+        }
+
+        .adressWrapper {
+          .street,
+          .city,
+          .phone,
+          .email {
+            font-size: 18px;
+            font-weight: 400;
+            margin-bottom: 8px;
+          }
+
+          .city {
+            margin-bottom: 40px;
+          }
+        }
+
+        .socialWrapper {
+          height: 48px;
+          display: flex;
+          justify-content: center;
+          margin-bottom: 8px;
+
+          .facebook,
+          .twitter {
+            margin-right: 48px;
+          }
+        }
+
+        .icon {
+          margin-top: 60px;
+        }
+      }
+
+      .contactForm {
+        width: 100%;
+        grid-area: right;
+        display: flex;
+        justify-content: stretch;
+
+        .formLabel {
+          display: block;
+          font-size: 18px;
+          text-transform: uppercase;
+          margin: 16px 0 8px;
+          color: $base-dark;
+          font-weight: 700;
+        }
+
+        .formTextInput {
+          height: 58px;
+          width: 100%;
+          padding: 16px 24px;
+          font-size: 16px;
+          letter-spacing: $medium-letter-spacing;
+          color: $base-dark;
+          flex-grow: 3;
+          border: 2px solid rgba(54, 72, 89, 0.1);
+          border-radius: 4px;
+
+          &::placeholder {
+            color: rgba(54, 72, 89, 0.15);
+            font-family: $base-font;
+            font-size: 18px;
+            letter-spacing: $medium-letter-spacing;
+            text-transform: uppercase;
+          }
+        }
+
+        .formTextArea {
+          @extend .formTextInput;
+          height: 215px;
+          resize: none;
+        }
+
+        .sendSection {
+          display: flex;
+          justify-content: stretch;
+          margin-top: 32px;
+          position: relative;
+          margin-bottom: 56px;
+
+          .rodoCheckbox {
+            position: absolute;
+            opacity: 0;
+            cursor: pointer;
+            height: 0;
+            width: 0;
+          }
+
+          .rodoCustomCheckbox {
+            position: absolute;
+            top: 0;
+            left: 0;
+            width: 24px;
+            height: 24px;
+            margin-right: 14px;
+            border: 2px solid rgba(54, 72, 89, 0.1);
+            border-radius: 1px;
+          }
+
+          .rodoCheckboxLabel {
+            padding-left: 38px;
+            font-size: 14px;
+            line-height: 21px;
+            font-weight: 300;
+            letter-spacing: $small-letter-spacing;
+          }
+        }
+        .formButton {
+          width: 100%;
+          @include button-primary;
+          height: 58px;
+          line-height: 58px;
+          border-radius: 4px;
+          font-size: 16px;
+          font-weight: 700;
+          padding: 0 16px;
+        }
+      }
+    }
+  }
+}
