+++ src/views/ContactOfficeView/ContactOfficeView.module.scss
@@ -1,0 +1,112 @@
+@import "../../variables.scss";
+@import "../../mixins.scss";
+
+.contactOfficeWrapper {
+  grid-area: hero;
+  .sectionTitle {
+    @include section-title;
+    margin: 140px 0 80px;
+    text-transform: uppercase;
+    text-align: left;
+  }
+
+  .sectionParagraph {
+    @include base-paragraph;
+  }
+
+  .sectionSubTitle {
+    @include section-sub-title;
+    font-size: 32px;
+    margin: 96px 0 24px;
+    text-align: left;
+  }
+
+  .salesOfficeBanner {
+    position: relative;
+    margin: 60px 0 160px;
+    width: 100%;
+    height: 234px;
+    display: grid;
+    grid-template-columns: 296px 296px auto;
+    grid-template-rows: auto;
+    gap: 0px 0px;
+    grid-template-areas: "left center right";
+    background: $base-grey;
+
+    .adressTextWrapper {
+      position: relative;
+      grid-area: left;
+      height: inherit;
+      display: flex;
+      flex-direction: column;
+      align-items: flex-start;
+      justify-content: center;
+      padding-left: 48px;
+
+      .adress,
+      .openingHours {
+        font-size: 20px;
+        font-weight: 700;
+        letter-spacing: $small-letter-spacing;
+        color: $base-dark;
+        margin: 0 0 8px 0;
+      }
+
+      .openingHours {
+        margin-top: 40px;
+      }
+
+      .city,
+      .hours {
+        font-size: 18px;
+        font-weight: 400;
+        letter-spacing: $small-letter-spacing;
+        color: $base-dark;
+        margin: 0;
+      }
+
+      &:after {
+        position: absolute;
+        content: "";
+        right: 0;
+        top: 16px;
+        height: 202px;
+        width: 2px;
+        border-radius: 2px;
+        background: $dark-grey;
+      }
+    }
+
+    .agentTextWrapper {
+      @extend .adressTextWrapper;
+      grid-area: center;
+
+      .name,
+      .contact {
+        @extend .adress;
+      }
+
+      .function,
+      .phoneNumber {
+        @extend .city;
+      }
+
+      .function {
+        text-transform: uppercase;
+      }
+      .contact {
+        @extend .openingHours;
+      }
+
+      &:after {
+        display: none;
+      }
+    }
+
+    .salesOfficeHexagons {
+      position: absolute;
+      right: 77px;
+      z-index: 2;
+    }
+  }
+}
